学校-人员管理 教师管理&&学生管理  CRUD操作

+++ SCE-DataAccess/src/main/java/cn/com/bonc/sce/dao/AccountDao.java
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Repository;
  * @since 2018/14/12 12:00
  */
 @Repository
-public interface AccountDao extends JpaRepository<Account, Integer> {
+public interface AccountDao extends JpaRepository<Account, Long> {
 
     Account getAccountByUserIdAndPassword( String userId, String password );
